#101 Move waitForPodStep.go to dogus package because it is only used by for dogus and not components anymore
