+++ app/setup/dogus/waitForPodStep.go
@@ -1,4 +1,4 @@
-package component
+package dogus
 
 import (
 	"context"
