+++ app/setup/dogus/waitForPodStep_test.go
@@ -1,4 +1,4 @@
-package component
+package dogus
 
 import (
 	"bytes"
